+++ tests/projects/package/multiconfig/test.lua
@@ -1,0 +1,7 @@
+-- main entry
+function main(t)
+
+    -- TODO
+    -- build project
+--    t:build()
+end